refactor(framework): Move Grid to flwr.serverapp
