+++ framework/py/flwr/compat/server/grid.py
@@ -15,7 +15,7 @@
 """Deprecated grid compatibility APIs."""
 
 
-from flwr.server.grid import Grid
+from flwr.serverapp import Grid
 
 
 class Driver(Grid):
