+++ framework/py/flwr/server/compat/app.py
@@ -23,8 +23,8 @@ from flwr.server.history import History
 from flwr.server.server import Server, init_defaults, run_fl
 from flwr.server.server_config import ServerConfig
 from flwr.server.strategy import Strategy
+from flwr.serverapp.grid import Grid
 
-from ..grid import Grid
 from .app_utils import start_update_client_manager_thread
