+++ framework/py/flwr/server/compat/app_utils.py
@@ -20,9 +20,9 @@ from collections.abc import Callable
 from typing import Any
 
 from flwr.common.typing import RunNotRunningException
+from flwr.serverapp.grid import Grid
 
 from ..client_manager import ClientManager
-from ..grid import Grid
 from .grid_client_proxy import GridClientProxy
