+++ framework/py/flwr/server/serverapp/app.py
@@ -59,7 +59,7 @@ from flwr.proto.appio_pb2 import (  # pylint: disable=E0611
     PullTaskInputResponse,
     PushTaskOutputRequest,
 )
-from flwr.server.grid.grpc_grid import GrpcGrid
+from flwr.superlink.grid import GrpcGrid
 from flwr.server.run_serverapp import run as run_
 from flwr.supercore.app_utils import start_parent_process_monitor
 from flwr.supercore.heartbeat import HeartbeatSender, make_task_heartbeat_fn_grpc
