+++ framework/py/flwr/serverapp/grid/__init__.py
@@ -12,15 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Public Flower ServerApp APIs."""
+"""Flower ServerApp Grid APIs."""
 
 
 from .grid import Grid
-from . import strategy
-from .server_app import ServerApp as ServerApp
 
 __all__ = [
     "Grid",
-    "ServerApp",
-    "strategy",
 ]
