+++ framework/py/flwr/serverapp/strategy/dp_adaptive_clipping.py
@@ -32,8 +32,9 @@ from flwr.common.differential_privacy import (
     compute_stdv,
 )
 from flwr.common.differential_privacy_constants import KEY_CLIPPING_NORM, KEY_NORM_BIT
-from flwr.server import Grid
-from flwr.serverapp.exception import AggregationError
+
+from ..exception import AggregationError
+from ..grid import Grid
 
 from .dp_fixed_clipping import validate_replies
 from .strategy import Strategy
