+++ framework/py/flwr/serverapp/strategy/fedavgm.py
@@ -33,7 +33,8 @@ from flwr.common import (
     RecordDict,
     log,
 )
-from flwr.server import Grid
+
+from ..grid import Grid
 
 from ..exception import AggregationError
 from .fedavg import FedAvg
