+++ framework/py/flwr/serverapp/strategy/fedopt.py
@@ -31,7 +31,8 @@ from flwr.common import (
     RecordDict,
     log,
 )
-from flwr.server import Grid
+
+from ..grid import Grid
 
 from ..exception import AggregationError
 from .fedavg import FedAvg
