+++ framework/py/flwr/serverapp/strategy/fedprox.py
@@ -29,7 +29,8 @@ from flwr.common import (
     RecordDict,
     log,
 )
-from flwr.server import Grid
+
+from ..grid import Grid
 
 from .fedavg import FedAvg
