+++ framework/py/flwr/serverapp/strategy/fedxgb_cyclic.py
@@ -28,7 +28,8 @@ from flwr.common import (
     RecordDict,
     log,
 )
-from flwr.server import Grid
+
+from ..grid import Grid
 
 from .fedavg import FedAvg
 from .strategy_utils import sample_nodes
