+++ framework/py/flwr/serverapp/strategy/strategy.py
@@ -22,7 +22,8 @@ from collections.abc import Callable, Iterable
 from logging import INFO
 
 from flwr.common import ArrayRecord, ConfigRecord, Message, MetricRecord, log
-from flwr.server import Grid
+
+from ..grid import Grid
 
 from .result import Result
 from .strategy_utils import log_strategy_start_info
