+++ framework/py/flwr/superlink/grid/inmemory_grid.py
@@ -27,7 +27,7 @@ from flwr.common.typing import Run
 from flwr.proto.node_pb2 import Node  # pylint: disable=E0611
 from flwr.server.superlink.linkstate import LinkStateFactory
 
-from .grid import Grid
+from flwr.serverapp.grid import Grid
 
 
 class InMemoryGrid(Grid):
